+++ statService/src/main/java/com/stat/dto/GameScoreDto.java
@@ -12,7 +12,7 @@ public class GameScoreDto {
 
 	String date;
 
-	Long gameId;	// 게임 아이디
+	int gameId;	// 게임 아이디
 	int score;	// 게임 점수
 	int endurance;	// 지구력
 	int resilience;	// 회복 탄력성
